Feat/revision 01

+++ GoGreen/Controllers/EcoViolationController.cs
@@ -110,15 +110,14 @@ namespace GoGreen.Controllers
 
         }
 
-
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult<EcoViolationResponse>> Put(int id, [FromBody] EcoViolationMunicipalityRequest request)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(id))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var type = await _context.EcoViolationStatuses.FindAsync(request.EcoViolationStatus.Id);
@@ -128,18 +127,6 @@ namespace GoGreen.Controllers
                 return BadRequest("Eco Violation Status not found");
             }
 
-            var user = await _context.User.Include(e => e.Municipality).FirstOrDefaultAsync(u => u.Id == userId);
-
-
-            var data = await _context.EcoViolations
-                        .Where(a => a.MunicipalityId == user.MunicipalityId)
-                        .Where(a => a.Id == id)
-                        .FirstOrDefaultAsync();
-
-            if (data == null)
-            {
-                return BadRequest("The user has no permission for this action");
-            }
 
             var updatedData = await _ecoViolationService.Update(id, request);
 
@@ -149,9 +136,15 @@ namespace GoGreen.Controllers
 
 
         // DELETE: api/EcoViolation/5
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+
+            if (!await CheckPermisionAsync(id))
+            {
+                return BadRequest("The user has no permission for this action");
+            }
 
             var isDeleted = await _ecoViolationService.Delete(id);
 
@@ -163,7 +156,30 @@ namespace GoGreen.Controllers
             return NoContent();
         }
 
+        private async Task<bool> CheckPermisionAsync(int id)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
+
+            if (string.IsNullOrEmpty(userId))
+            {
+                return false;
+            }
+
+            var user = await _context.User.Include(e => e.Municipality).FirstOrDefaultAsync(u => u.Id == userId);
+            var data = await _context.EcoViolations
+                        .Where(a => a.MunicipalityId == user.MunicipalityId)
+                        .Where(a => a.Id == id)
+                        .FirstOrDefaultAsync();
+
+            if (data == null)
+            {
+                return false;
+            }
+
+            return true;
+
+        }
 
     }
 }
